backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.6.1
python-dotenv==1.0.1
email-validator==2.1.0.post1
openai
alembic==1.13.1
psycopg2-binary==2.9.9
